Split prettier code off to a separate module
